requirements.txt
Flask>=1.0.3
Flask-SQLAlchemy>=2.4.0
python-dotenv>=0.10.3
SQLAlchemy>=1.3.4